pkg for wasm module
