+++ client/devfile-builder/wasm/pkg/global/global.go
@@ -1,0 +1,9 @@
+package global
+
+import (
+	"github.com/devfile/library/v2/pkg/devfile/parser"
+	"github.com/devfile/library/v2/pkg/testingutil/filesystem"
+)
+
+var Devfile parser.DevfileObj
+var FS filesystem.Filesystem
